Create code
This code uses the request library to make a GET request to the World Bank's API to retrieve the inflation rate data for China. The data is returned in JSON format, and the inflation rate is accessed and logged to the console.
Note: this code snippet doesn't include the process of installing the 'request' library you need to install it in your project before running this code.